+++ src/Renderer/Buffer.h
@@ -15,5 +15,5 @@ class IndexBuffer {
         virtual ~IndexBuffer() {}
         virtual void Bind() const = 0;
         virtual void Unbind() const = 0;
-        static IndexBuffer* Create(float* indices, uint32_t size);
+        static IndexBuffer* Create(uint32_t* indices, uint32_t size);
 };
